# server/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3001",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/professionals/add": {
      "post": {
        "tags": [
          "Professionals"
        ],
        "description": "Add new professional",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Add new user",
            "schema": {
              "type": "object",
              "properties": {
                "description": {
                  "type": "string",
                  "example": "description"
                },
                "locations": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "Consultorio Lo Zozzo"
                      },
                      "address": {
                        "type": "string",
                        "example": "Via i laziali da Roma, 42"
                      },
                      "lat": {
                        "type": "number",
                        "example": 41.9357261
                      },
                      "lng": {
                        "type": "number",
                        "example": 12.4839667
                      }
                    }
                  }
                },
                "phone": {
                  "type": "string",
                  "example": "892"
                },
                "userId": {
                  "type": "string",
                  "example": ""
                }
              },
              "required": [
                "description",
                "phone"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/professionals/": {
      "get": {
        "tags": [
          "Professionals"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/professionals/id/{id}": {
      "get": {
        "tags": [
          "Professionals"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id of a professional"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/professionals/distance": {
      "get": {
        "tags": [
          "Professionals"
        ],
        "description": "",
        "parameters": [
          {
            "name": "latitude",
            "in": "body",
            "description": "Latitude",
            "required": true,
            "type": "number"
          },
          {
            "name": "longitude",
            "in": "body",
            "description": "Longitude",
            "required": true,
            "type": "number"
          },
          {
            "name": "latitude\r",
            "in": "query"
          },
          {
            "name": "longitude\r",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/users/add": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Add new User",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Add new user",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Mario"
                },
                "surname": {
                  "type": "string",
                  "example": "bianchi"
                },
                "mail": {
                  "type": "string",
                  "example": "mariooo@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "iammariobianchi"
                }
              },
              "required": [
                "name",
                "surname",
                "mail",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Get all Users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/users/id/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Get single user by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id of a user"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {}
}